abstract processing xlsx file into own class
